CI: do not "fail" branches upon coverage decrease
Right now the project homepage on GitHub shows an red failure due to
[1]. While I agree that coverage reports are a useful metric, they are
most relevant when considering patches (or changes) on their own. I do
not think it makes any sense to show them on branches. When we merged
that commit, the reviewer had the info about the coverage change, and we
just do not have 100% coverage. Let's not pretend that that is a
blocker.

[1] https://codecov.io/gh/Telecominfraproject/oopt-gnpy/commit/924c56850d77c53f8e5c46e21082e5510a8b00d5

Change-Id: I652ada3357e521c9a3351faac2f9c2e8e4aa5773