tests: activate mccabe the complexity checker
It turns out that my text editor has been warning about "excessive code
complexity" via the python-language-server plugin that uses pycodestyle
underneath. As the test suite is configured to use `flake8` instead,
let's make sure that the results are consistent, and configure the code
complexity threshold for the CI linter run as well.

I'm not sure if this is usable when run incrementally, though (such as
in the CI). I *think* that `flake8 --diff` disregards warnings issued on
lines which have not been changed in the corresponding diff, while
mccabe reports its warnings for the line with the function signature --
and that one is not included in the output of `git diff -U0`. As a
result, I suspect that if a patch increases the code complexity of a
function over the threshold, our CI setup might not report that.

However, a non-diff run of the linters via `flake8` will now start
reporting C901 warnings. For those using `python-language-server` to
integrate with their IDEs and text editors, install `pyls-flake8` to
preserve the existing functionality.

Change-Id: If02e7f189514077aca3e865e1882a4a3fabb4dba