Fix z position array creation within Raman solver
The Raman solver gave the wrong loss profile when the fiber length was
not a multiple of the simulation parameter space_resolution as,
in this case, the fiber termination position was not included within
the considered z position array and the output loss profile was
evaluated at the wrong position.

Additionally, the Raman solver failed when the simulation parameter
space_resolution was greater than the fiber length as the z position
array contained only one element.

With this version the fiber termination position is always included
in the z position array and is composed of multiple uniformly spaced
positions and the final position (in general, the latter is not
uniformly spaced).

Example:
* fiber_length = 5, space_resolution = 4 => old version z = [0, 4]
                                          => new version z = [0, 4, 5]

* fiber_length = 5, space_resolution = 6 => old version z = [0]
                                          => new version z = [0, 5]

Alternative:

z = linespace(0, fiber_length, int(ceil(fiber_length/space_resolution)))

PROS
* Solves the first issue
* Returns a uniformly spaced z position array (there is not a
  straightforward advantage in the simulation)

CONS
* Does not solve the second issue
* It is slightly more involved

Change-Id: I8886c3563ac7305c49cb5915712777ef561c5d4f