Fix behavior when there's no EDFA DGT
We ship "some" DGT values which effectively mask this, but it's possible
to provide a "trivial" DGT vector such as [0, 0]. When that happens, the
code was failing with a numpy-level warning related to a division by
zero.

The code tried to be ready for this by trying to catch an exception, but
this relied on a particular numpy behavior upon zero division which was
not set up properly. With a correct numpy setting, the exception thrown
is a different one, so adapt for that as well.

Note that there are two possible cases of division by zero:

- on a zero tilt, it's a case of `0.0 / np.float64(0.)`, which is
controlled via "invalid",
- on a non-zero tilt, it's a case of `<float> / np.float64(0.)`, which
leads to "divide".

Change-Id: I7a3c0b6b9b85e90a8a7201b7c7a56a5a44c27d69