YANG: topology info for ONOS
These bits are irrelevant for GNPy, but required on the ONOS side. I'm
not particularly happy with the format of the links, or the fact that we
need them to be done manually. I really wanted to generate them from the
underlying GNPy topology, however, that's hard because ONOS needs port
information (so that it knows how to configure these ROADMs).

FIXME: It might be possible if everything is in the same network
instance. In that case, this model would just augment each link with
port IDs.  However, I'm not really happy with having "unrecognized"
elements in the GNPy-specific topology. Or perhaps that's actually OK?

Change-Id: I31be96c04f069ab797b9da82993633dc804180e2