Take explicitly set out_voa into account when adding padding
When the loss of a span is smaller than a padding value specified in
the eqpt file, attenuation (att_in) is added at the span input by
auto-design. This happens even if the user has already explicitly set
attenuation at the output of the Edfa (out_voa) before the span, i.e.
out_voa is not considered when calculating span loss. With this patch,
the out_voa is taken into account in the span loss and padding
calculation.

Signed-off-by: Jonas Mårtensson <jonas.martensson@ri.se>
Change-Id: I57dbe666ef50738dd7465d4fda61e53280823090