Adapt loss and ref power computation
Loss is now computed based on signal power to account for Raman gain
during pref update.
Moreover, computation of p_spani was based on average channel power. This
however can not hold in the case of mixed rates and mixed per channel power.
This commit changes this computation using the loss of Fused, the loss of
Fiber, the gain of Edfa and the initial refeference p_spani.

This corrects some of the inconsistencies when printing element powers,
because of the integration of noise or not in p_spani (eg in 505386
"Raman Solver restructuring and speed up" patchset 13)

Signed-off-by: EstherLerouzic <esther.lerouzic@orange.com>
Change-Id: I5392f906061400834916915199aed10e2ac5553c