Change pref from a scalar to a list of per channel delta power
The idea behind this change is to reproduce the exact same behaviour as with
the scalar, but accounting for variable levels of powers.

- delete the  neq_ch: equivalent channel count in dB because with mixt
rates and power such a value has limited utility

- instead creates a vector that records the 'user defined' distribution
of power.
This vector is used as a reference for channel equalization out of the ROADM.
eg if user specifies freq1: 1dBm, freq2: +3dBm, freq3:-3dBm, recorded
in p_span0_per_channel and if power for the reference channel is 0dBm
(recorded in p_span0) , the same relative power difference will be applied
on each degree of ROADMs.
target_pch_out_db is the target power for the reference channel so
p_out_target[f1, f2, f3] = target_pch_out_db + p_span0_per_channel[f1, f2, f3] -p_span0
                         = -19, -16, -22

- in amplifier: the saturation is computed based on this p_span0_per_channel
instead of the nb of channels. The target of the future refactor will be to
use the effective carrier's power.
I prefer to have this first step, because this is how it is done today,
and I would like first to add more behaviour tests before doing this refactor.

- in spectralInfo class, change pref to a Pref object to ease si operations
to enable __add__ operation also on p_span0_per_channel.

Signed-off-by: EstherLerouzic <esther.lerouzic@orange.com>
Change-Id: Ifacea50e77a2eb4306d80c640e5277b491185488

Signed-off-by: EstherLerouzic <esther.lerouzic@orange.com>
Change-Id: I591027cdd08e89098330c7d77d6f50212f4d4724