Add a user defined initial spectrum in propagation functions
A new function is added to use requests specs based on transceivers
and on user definition of mixed rates and powers.

Propagation functions are changed so that, if the user defines a
specific distribution, then they use it else they use as before,
all identical channels based on the initial request. In this case,
as before this change, we assume full load, with same channel for
the spectral info and not the resulting mixt of channels after
routing.

Signed-off-by: EstherLerouzic <esther.lerouzic@orange.com>
Change-Id: Icf56396837b77009e98accd27fcebd2dded6d112