Add the reference_carrier in the spectrum information object
ref_carrier conveys additional information about baud rate that was used
for the design (would it be with auto-design of for a given design).
Other attributes (like roll-off, slot_width) may be added there for
future equalization purpose.
This reference baud_rate is required to compute reference target power
based on spectral density values during propagation. It is thus required
because of the on the fly evaluation of loss for p_span_i and for printing
loss and target power of ROADM during propagation.

Signed-off-by: EstherLerouzic <esther.lerouzic@orange.com>
Change-Id: Ic7441afa12ca5273ff99dea0268e439276107257